Feat: 뮤멘트 상세보기 뷰 전체 UI#24
